Made the script more efficient
Create a csv file containing paths to the covariates I'd like to use in
the model, and the names that I'd like to re-assign to the covariates. I
can read in the csv, and create a list of the paths to loop through and
open; this removes the need to have 100 rows of code listing the file
paths, and another 100 rows listing the new names I'd like to assign to
the file.